requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-client==0.0.1
mysqlclient==2.0.3
python-dotenv==0.19.1
requests==2.26.0
SQLAlchemy==1.4.26
urllib3==1.26.7
Werkzeug==2.0.2
